+++ app.go
@@ -20,7 +20,7 @@ import (
 	"github.com/wailsapp/wails/v2/pkg/runtime"
 )
 
-var version = "0.3.3-alpha"
+var version = "0.4.0-alpha"
 
 type Config = config.Config
 type GameInfo = game.GameInfo